# OpenAPI.yaml
openapi: "3.0.3"
info:
  title: Bangumi Recorder API
  version: "2.1.0"
  description: |
    自托管的番剧追番记录器 API v2。

    所有 v2 接口统一返回格式：
    ```json
    { "status": 0, "data": ..., "message": "" }
    ```
    - `status`: `0` 成功 / `-1` 报错
    - `data`: 所有响应数据
    - `message`: 错误描述（成功时为空）

    错误时返回适当的 HTTP 状态码（400 / 404 / 409 / 401 / 500 等）。
servers:
  - url: ""
    description: 与前端同源
paths:
...
  /api/v2/auth/login:
    post:
      summary: 登录
      tags: [认证]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
              required: [username, password]
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/LoginData"
        "401":
          description: 用户名或密码错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/v2/auth/register:
    post:
      summary: 注册
      tags: [认证]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
                register_token: { type: string, description: "注册令牌（如配置需要）" }
              required: [username, password]
      responses:
        "200":
          description: 注册成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/RegisterData"
        "403":
          description: 注册已关闭
        "409":
          description: 用户名已存在
...
  /api/v2/me:
    get:
      summary: 获取当前用户信息
      tags: [用户]
      security: [{ BearerJWT: [] }]
      responses:
        "200":
          description: 用户信息
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/UserInfo"
    patch:
      summary: 更新用户资料
      tags: [用户]
      security: [{ BearerJWT: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname: { type: string }
                avatar: { type: string }
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiOk"
  /api/v2/me/password:
    put:
      summary: 修改密码
      tags: [用户]
      security: [{ BearerJWT: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                old_password: { type: string }
                new_password: { type: string, minLength: 6 }
              required: [old_password, new_password]
      responses:
        "200":
          description: 密码修改成功（之后需重新登录）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiOk"
        "400":
          description: 原密码错误等
  /api/v2/me/token:
    post:
      summary: 重新生成 API Token
      description: 创建一个新的 Token（全部权限），旧的 Token 仍然有效
      tags: [用户]
      security: [{ BearerJWT: [] }]
      responses:
        "200":
          description: 新生成的 Token
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/TokenData"

  # ────────────────────────────── API Token 管理 ─────────────────────
  /api/v2/tokens:
    get:
      summary: 列出所有 API Token
      tags: [API Token 管理]
      security: [{ BearerJWT: [] }]
      responses:
        "200":
          description: Token 列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ApiTokenItem"
    post:
      summary: 创建新的 API Token
      tags: [API Token 管理]
      security: [{ BearerJWT: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, permissions]
              properties:
                name:
                  type: string
                  description: Token 名称
                permissions:
                  type: integer
                  description: 权限 bitmask
      responses:
        "200":
          description: 创建成功，返回原始 Token（仅此一次）
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/CreateTokenData"
  /api/v2/tokens/{id}:
    put:
      summary: 更新 API Token
      tags: [API Token 管理]
      security: [{ BearerJWT: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                permissions: { type: integer }
                is_active: { type: boolean }
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiOk"
        "404":
          description: Token 未找到
    delete:
      summary: 删除/吊销 API Token
      tags: [API Token 管理]
      security: [{ BearerJWT: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiOk"
        "404":
          description: Token 未找到
  /api/v2/tokens/permissions:
    get:
      summary: 获取权限标签列表
      tags: [API Token 管理]
      security: [{ BearerJWT: [] }]
      responses:
        "200":
          description: 权限标签列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PermissionLabel"

  # ────────────────────────────── 搜索 ──────────────────────────────
  /api/v2/search:
    get:
      summary: 在线搜索 Bangumi 条目
      description: 搜索 bgm.tv，结果同时存入本地缓存。
      tags: [搜索]
      security: [{ BearerJWT: [] }]
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
          description: 搜索关键词
        - name: page
          in: query
          schema: { type: integer, default: 1 }
      responses:
        "200":
          description: 搜索结果列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/BangumiSearchItem"
        "400":
          description: 缺少搜索关键词
  /api/v2/bangumi/{id}:
    get:
      summary: 获取 Bangumi 条目详情
      description: |
        优先返回本地缓存（24h 内），缓存过期或 `?force=true` 时从 bgm.tv 抓取。
      tags: [搜索]
      security: [{ BearerJWT: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
          description: Bangumi 外部 ID
        - name: force
          in: query
          schema: { type: boolean, default: false }
          description: 强制刷新缓存
      responses:
        "200":
          description: 条目详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/BangumiItem"
        "404":
          description: 条目未找到
  /api/v2/bangumi/{id}/refresh:
    post:
      summary: 强制刷新 Bangumi 条目内容
      description: 跳过 24 小时缓存，重新抓取并覆盖标题、类型、封面与详情缓存。
      tags: [搜索]
      security: [{ BearerJWT: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
          description: Bangumi 外部 ID
      responses:
        "200":
          description: 已刷新的条目详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/BangumiItem"
        "404":
          description: 条目未找到
        "500":
          description: 上游抓取或缓存写入失败
  /api/v2/bangumi/{id}/episodes:
    get:
      summary: 获取剧集元数据列表（爬取 bgm.tv 并缓存，支持 ?force=true 强制刷新）
      tags: [搜索]
      security: [{ BearerJWT: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
          description: Bangumi 外部 ID
        - name: force
          in: query
          schema: { type: boolean, default: false }
          description: 强制从 bgm.tv 重新爬取，跳过缓存
      responses:
        "200":
          description: 剧集元数据列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/BangumiEpisodeMeta"
  /api/v2/imdb/search:
    get:
      summary: 在线搜索 IMDb 条目
      description: |
        支持免 API 的 IMDb suggestion 接口；配置 `OMDB_API_KEY` / `IMDB_API_KEY` 后可通过 `use_api=true` 使用 OMDb API。
        搜索结果会写入 `external_media` 本地缓存。
      tags: [搜索]
      security: [{ BearerJWT: [] }]
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
          description: 搜索关键词
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: use_api
          in: query
          schema: { type: boolean, default: false }
          description: 是否优先使用 OMDb API
      responses:
        "200":
          description: IMDb 搜索结果列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ImdbSearchItem"
        "400":
          description: 缺少搜索关键词
  /api/v2/imdb/{id}:
    get:
      summary: 获取 IMDb 条目详情
      description: |
        按 IMDb `tt...` ID 获取详情，优先使用 24h 本地缓存；`force=true` 可强制刷新。
      tags: [搜索]
      security: [{ BearerJWT: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, example: tt0816692 }
          description: IMDb title ID 或 IMDb title URL
        - name: force
          in: query
          schema: { type: boolean, default: false }
          description: 强制刷新缓存
        - name: use_api
          in: query
          schema: { type: boolean, default: false }
          description: 是否优先使用 OMDb API
      responses:
        "200":
          description: IMDb 条目详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ImdbItem"
        "404":
          description: 条目未找到
  /api/v2/search/local:
    get:
      summary: 搜索本地缓存
      tags: [搜索]
      security: [{ BearerJWT: [] }]
      parameters:
        - name: q
          in: query
          schema: { type: string }
          description: 关键词
        - name: id
          in: query
          schema: { type: integer }
          description: 按 ID 精确搜索
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        "200":
          description: 本地搜索结果
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/LocalSearchResult"

  # ────────────────────────────── 追番记录 ──────────────────────────────
  /api/v2/records:
    get:
      summary: 列出所有追番记录
      tags: [追番记录]
      security: [{ BearerJWT: [] }]
      responses:
        "200":
          description: 追番列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/RecorderItem"
    post:
      summary: 添加追番记录
      tags: [追番记录]
      security: [{ BearerJWT: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddRecordParams"
      responses:
        "200":
          description: 添加成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/AddRecordData"
        "400":
          description: 参数无效
        "404":
          description: Bangumi 未找到
        "409":
          description: 记录已存在
  /api/v2/records/detail:
    get:
      summary: 列出所有追番记录（含详细条目信息）
      tags: [追番记录]
      security: [{ BearerJWT: [] }]
      responses:
        "200":
          description: 详细追番列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/DetailListItem"
  /api/v2/records/bangumi/{id}:
    get:
      summary: 按 Bangumi ID 获取追番进度
      tags: [追番记录]
      security: [{ BearerJWT: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 追番进度
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/GetRecordData"
    patch:
      summary: 按 Bangumi ID 更新追番进度
      tags: [追番记录]
      security: [{ BearerJWT: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                recorder: { type: string, description: "进度，格式 集数|分钟:秒" }
                user_status:
                  type: integer
                  enum: [0, 1, 2, 3, 4]
                  description: "0=想看 1=在看 2=看过 3=搁置 4=抛弃"
      responses:
        "200": { description: 更新成功, content: { application/json: { schema: { $ref: "#/components/schemas/ApiOk" } } } }
        "400": { description: 参数缺失 }
        "404": { description: 记录未找到 }
    delete:
      summary: 按 Bangumi ID 删除追番记录
      tags: [追番记录]
      security: [{ BearerJWT: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: hard_delete
          in: query
          schema: { type: boolean, default: false }
          description: false=软删除，true=硬删除 recordings 记录
      responses:
        "200": { description: 删除成功, content: { application/json: { schema: { $ref: "#/components/schemas/ApiOk" } } } }
        "404": { description: 记录未找到 }
...
  /api/v2/open/other/{id}:
    get:
      summary: 获取自定义条目详情（API Token 鉴权）
      tags: [开放接口]
      security:
        - BearerApiToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: token
          in: query
          required: false
          description: "旧客户端兼容参数；推荐使用 Authorization: Bearer <api-token>。"
          schema: { type: string }
      responses:
        "200":
          description: 自定义条目详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...
components:
  securitySchemes:
    BearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 在 /api/v2/auth/login 获取 Token，以 `Bearer <token>` 形式传入 Authorization 头。
    BearerApiToken:
      type: http
      scheme: bearer
      bearerFormat: API Token
      description: 在 /api/v2/tokens 创建 Token，以 `Bearer <api-token>` 形式传入 Authorization 头。旧接口的 token 查询参数仅为兼容保留。
  schemas:
...
    ApiResponse:
      type: object
      properties:
        status:
          type: integer
          enum: [0, -1]
          description: "0=成功, -1=报错"
        data: {}
        message:
          type: string
          nullable: true
    ApiOk:
      type: object
      properties:
        status: { type: integer, enum: [0] }
        data: { nullable: true, default: null }
        message: { type: string, nullable: true }
    ApiError:
      type: object
      properties:
        status: { type: integer, enum: [-1] }
        data: { nullable: true, default: null }
        message: { type: string }

    # ── 认证 ──
    LoginData:
      type: object
      properties:
        token: { type: string }
      required: [token]
    RegisterData:
      type: object
      properties:
        token: { type: string }
        api_token: { type: string }
      required: [token, api_token]
...